GW-72: Refactored grid grow into shared component
