+++ client/src/pages/rantangan/components/constants.js
@@ -6,10 +6,12 @@ const PAST_RANTANGAN = {
 		{
 			image: RANTANGAN_22_23[0],
 			title: 'Nasi Liwet',
+			description: 'Friday, March 17 2023',
 		},
 		{
 			image: RANTANGAN_22_23[1],
 			title: 'Sate Ayam Sambal Matah',
+			description: 'Sunday, May 29 2022',
 		},
 	],
 	'2021/2022': [],
