+++ client/src/shared/data/upcoming_rantangan.js
@@ -4,10 +4,16 @@ const UPCOMING_RANTANGAN = [
 	{
 		image: rantangan,
 		title: 'Title of Rantangan',
-		date: 'Friday, August 4 | 8-9:40 PM',
-		loc: 'Online on Zoom',
-		link: 'forms.gle/AXqcbDGtKBBiHCqX6',
+		description: 'upcoming',
 	},
 ];
 
 export default UPCOMING_RANTANGAN;
+
+// {
+//	image: rantangan,
+//	title: '',
+//	description: '',
+//  location:'',
+//  price: '',
+// },
